Move file system commands to new folder
