+++ src/client.ts
@@ -48,7 +48,7 @@ import BatteryStatusCommand from './commands/host-transport/batteryStatus';
 import ClearCommand from './commands/host-transport/clear';
 import Connect from './commands/host/connect';
 import { Connection } from './connection';
-import CpCommand from './commands/host-transport/cp';
+import CpCommand from './commands/host-transport/fileSystem/cp';
 import Disconnect from './commands/host/disconnect';
 import FileStatCommand from './commands/host-transport/fileStat';
 import { FileStat } from './filestats';
@@ -70,10 +70,10 @@ import ListReversesCommand from './commands/host-transport/listreverses';
 import ListSettingsCommand from './commands/host-transport/listSettings';
 import LogcatCommand from './commands/host-transport/logcat';
 import { LogcatReader } from './logcat/reader';
-import MkDirCommand from './commands/host-transport/mkdir';
+import MkDirCommand from './commands/host-transport/fileSystem/mkdir';
 import { Monkey } from './monkey/client';
 import MonkeyCommand from './commands/host-transport/monkey';
-import MvCommand from './commands/host-transport/mv';
+import MvCommand from './commands/host-transport/fileSystem/mv';
 import { PullTransfer } from './sync/pulltransfer';
 import { PushTransfer } from './sync/pushtransfer';
 import PutSetting from './commands/host-transport/putSetting';
@@ -81,7 +81,7 @@ import { Readable } from 'stream';
 import RebootCommand from './commands/host-transport/reboot';
 import RemountCommand from './commands/host-transport/remount';
 import ReverseCommand from './commands/host-transport/reverse';
-import RmCommand from './commands/host-transport/rm';
+import RmCommand from './commands/host-transport/fileSystem/rm';
 import RootCommand from './commands/host-transport/root';
 import ScreenShotCommand from './commands/host-transport/screencap';
 import SetProp from './commands/host-transport/setProperty';
@@ -94,7 +94,7 @@ import SyncCommand from './commands/host-transport/sync';
 import SyncEntry from './sync/entry';
 import TcpCommand from './commands/host-transport/tcp';
 import TcpIpCommand from './commands/host-transport/tcpip';
-import TouchCommand from './commands/host-transport/touch';
+import TouchCommand from './commands/host-transport/fileSystem/touch';
 import TrackCommand from './commands/host/trackdevices';
 import { Tracker } from './tracker';
 import UninstallCommand from './commands/host-transport/uninstall';
